UAVcrop/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision
numpy
opencv-python
pyyaml
